Extract forwarder push logic to separate file
This change should help to achieve better
structure and readability.
